remove omni from url
